Combine CSV/TSV import tests with separator tests
